backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pillow==10.4.0
moviepy==1.0.3
python-multipart==0.0.9
pytest==8.3.2
httpx==0.27.2
openai
elevenlabs
requests
python-dotenv==1.0.1
gTTS==2.5.1
openai-whisper==20231117
deepgram-sdk==3.4.2